Added JScrollBar containing a JTree which contains the declarations.
declarations contain: parameters, types, variables and functions.
 this class can add, remove these functions.

JTree is a list representation of the declarations. The JTree has theability to toggle the 4 different types of declarations. The JTree is initialized with a declarations pointer

JScrollBar is a container for the JTree. Allows for an easier access of the different declarations when the list becomes larger.

Currently, we create these 3 classes in the startup class.